Scope skjema-oppvarming til skjema-tester som blir eget setup-prosjekt
Oppvarmingen lå i global-setup.ts og ble brukt på HVER kjøring, også når
ingen skjema-tester kjørte

Oppvarmingen flyttes til et Playwright setup-prosjekt (tests/skjema/skjema.setup.ts) som
skjema-spec-ene avhenger av via dependencies: ['skjema-setup']. Da kjøres
oppvarmingen kun når skjema-tester faktisk er valgt, og kun én gang.
